pyproject.toml
[project]
name = "dave-ledger"
version = "0.1.0"
description = "DAVE Ledger: Dynasty Asset Valuation Engine Ledger"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jupyterlab>=4.5.1",
    "nflreadpy>=0.1.5",
    "numpy>=1.26.4",
    "pandas>=1.5.3",
    "pyarrow>=22.0.0",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
